requirements.txt
lark-parser==0.8.2
oddhash==0.0.1
pkg-resources==0.0.0
pycryptodome==3.9.7
